rename auth into auth_service
