--- servicelog.api/auth/authen.py
@@ -1,31 +1,0 @@
-from datetime import datetime, timedelta
-from typing import Optional
-from fastapi import Depends, HTTPException, status
-from fastapi.security import OAuth2PasswordBearer
-from jose import JWTError, jwt
-from schemas.user_schema import UserResponse
-from sqlalchemy.orm import Session
-from schemas.login_schema import LoginResponse
-from services.user_service import get_user
-
-# Secret Key and Algorithm
-SECRET_KEY = "This-is-a-secret"
-ALGORITHM = "HS256"
-ACCESS_TOKEN_EXPIRE_MINUTES = 30
-
-# OAuth2 Password Flow
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="auth/token")
-
-def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
-    to_encode = data.copy()
-    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES))
-    to_encode.update({"exp": expire})
-    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-
-
-def decode_access_token(token: str):
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        return payload
-    except JWTError:
-        return None
